requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
anthropic==0.39.0
streamlit==1.40.0
pandas==2.2.0
python-dotenv==1.0.0
